Support declaring resources as read-only.
This enables referencing to resources that we do not want Deployster to touch or modify in a safe way. If those resources are MISSING or STALE, Deployster will abort the plan instead of creating or updating them, but as long as they are VALID, other resources that reference them will receive them as their dependencies.