Introduce "GcpServices" class.
This class provides a GCP services facade. The main incentive to create this class is to enable unit-testing, but it's also a more correct form of encapsulation of GCP services (vs. module methods as it was until now in "gcp_services.py").

Adapt "GcpResource" and "GcpProject" to use this new class.

Add tests of "GcpProject" accordingly.