Refactor GKE cluster resource to next Deployster version.
This refactor adapts GKE cluster resource to the dependency refactoring this branch is progressing towards. The change is essentially removing direct dependencies between resources, instead requiring the deployer (user) to wire the dependencies via expressions in a resource's configuration clause.

So instead, for instance, that a GKE cluster requests a GCP project dependency, it instead requests a project ID (string); the user is in charge of providing that project ID - either manually (hardcoded string) or an expression that references another resource (indirect dependency).

This commit also adds the first GKE cluster test, using the new testing framework (py.test).