Refactor dependency mechanism (#37)
Next version of Deployster. Major themes:

### Resource dependencies

The resource dependency mechanism has been rewritten, and some core functionality has been changed. The main change is the drop of _Declarative dependencies_ from resource types. This means that resource implementation can no longer declare that they require other resources; in fact, resources will no longer receive resource dependencies in their JSON. It is still possible for the deployer to specify dependencies for a resource, but for the single purpose of allowing them to be used in expressions under the `config` section of the resource.

This means that, for example, in order to pass a GCP Compute IP address' actual IP to a GKE cluster resource, you would specify the IP address resource as a dependency of the cluster resource, and in the cluster's `config` section, you would use the expression `{{ my_ip.actual_ip }}` to pass the IP address to the cluster - just as any other configuration value!

### Plan removal

Deployster will no longer print a plan before execution. This means that you will not know in advance the plan that will be executed, until it's actually executed. The reason for this change is that it is not possible to anticipate in advance the exact set of actions that will be required for some resources, especially if a resource does not exist, and other resources depend on it. 

To mitigate this change, a new flag has been added: `--confirm=NO|ONCE|RESOURCE|ACTION` which allows the deployer to request confirmation at certain events:
- `NO` means run on auto-pilot (suitable for continuous delivery, for instance)
- `ONCE` means one time at the beginning of the run
- `RESOURCE` means before every resource
- `ACTION` means before every action to be taken (this is the default)

### Resource protocol changes

The following changes have been made to the resource protocol:

- resource state can now only be `VALID` or `STALE` (where `STALE` also includes the old `MISSING` state)
- resource configuration must be an object (it can no longer be an array)

### Other changes

- Deployster now takes paths from the host machine. This means that referneces to paths in manifests can use the host machine's paths (instead of using the `_work` & `_workspace` builtin variables)
- Internal refactoring to improve code maintainability and testability
- Finally added testing with a (current) coverage of over 70% (using the excellent PyTest framework and Coveralls.io)