Miscellaneous fixes (#52)
* Fix key-error exception in Kubernetes resources.

* Minor refactor of "k8s_main.py"

* Override of "state()" method needs to re-annotate itself with "@action".

This is required so the "state" action is still recognized by the "execute(..)" method.
Adapt testing of "check_availability" accordingly (validation of "timeout_ms > timeout_interval_ms" was moved from the constructor to the "state" method override)

* Print expression context when expression in manifests fails.

* Bug fixes.

Fix paths sending from "deployster.py" to the "Manifest" constructor.
Always work on absolute paths when mounting plugs to resource images.
When confirmation mode is "RESOURCE", only ask for confirmation if the resource actually has actions to perform.
Fix incorrect state bug.

* Fail when target GCP project is pending deletion, or is not active.

* Disable YUM "fastestmirror" plugin in CentOS images, and also install "which" package.

* Print Kubernetes objects being created/deleted when in verbose mode, and infer their "apiVersion" & "kind" properties when missing from the manifest.

* Added TODO for caching "init" actions.

* Don't fail if 'lifecycleState' not present in GCP project state (mainly for tests, but good to be on the safe side anyway).

* Fix tets.

* Added TODO to "manifest.py"

* Fix swallowed exception in "ExternalServices.create_gcp_sql_instance"

* Support verbosity in creating Cloud SQL instances.

Also fix tests not to send "config" to "init" action of resources.

* Fix exception caused when displaying configuration validation errors.

* Improve caching of resource images.

* Removed implemented TODOs and those that were converted to GitHub issues.

* Add support for creating Cloud SQL users.

This enables adding the "users" property to the configuration of a Cloud SQL resource. The resource will then make sure these users exist - it will not, however, update the passwords if they do exist (as we can't discover their current passwords to compare with the configuration).

* Fix Cloud SQL resource tests after adding support for "users" configuration property.

* Cloud SQL user name should not require a minimum lenght of 4

* Support Jinja '{% .. %}' expressions in manifests

* Support "include" operation in Jinja expressions.

* Match GCP project ID validation pattern to Google's rules.

* Add GCP Service Account resource support.

This enables the "infolinks/deployster-gcp-service-account" resource type.

* Rename "GcpServiceAccount" to "GcpIamServiceAccount".

This also renames the corresponding Dockerfile and Python script, along with updates to the relevant tests.

* Update GCP IAM Service Account resource to receive an Email address in the configuration clause.

This is required because there are different Email suffixes for different service account types, and only the user (deployer) can know the appropriate one.

* GCP Docker image no longer needs to copy DResource files, since sub-images copy those instead.

* Test image version should default to "0.0.0" when "BUILDKITE_COMMIT" is not defined.

* Optimize imports in "util.py"

* Test "util.post_process(..)" with a list containing strings with Jinja templates

* Add GCP IAM Policy resource.

This enables declaring that certain users or service accounts are granted permissions in a GCP project IAM policy.

* Print full IAM policy bindings before updating (for reference).

Since IAM policy bindings are so sensitive, it's imperative that the user has full knowledge of what happened, and how to revert. This is a step in this direction, by printing the IAM policy before & after the update.

* Finish SQL users implementation, and support SQL scripts post-processing.

* Fix README

* Improve startup greeting

* Support custom context for post-processed SQL scripts.